adding size_t in strlen

+++ 0x15-file_io/holberton.h
@@ -10,6 +10,6 @@
 
 int create_file(const char *filename, char *text_content);
 ssize_t read_textfile(const char *filename, size_t letters);
-int _strlen(char *s);
+ssize_t _strlen(char *s);
 
 #endif
